+++ features/public-form/use-cases/__tests__/load-public-survey-page.use-case.test.ts
@@ -4,7 +4,8 @@ import { getPublicFormAccessUseCase } from "@/features/public-form/use-cases/get
 import { loadPublicSurveyPageUseCase } from "@/features/public-form/use-cases/load-public-survey-page.use-case";
 import { getSubmissionByAccessTokenUseCase } from "@/features/public-submissions/edit/get-submission-by-access-token.use-case";
 import { ApiResult, type Submission } from "@/lib/endatix-api";
-import { Result } from "@/lib/result";
+import { ERROR_CODE } from "@/lib/endatix-api/shared/error-codes";
+import { Result, type ResultType } from "@/lib/result";
 import type { ActiveDefinition } from "@/types";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
@@ -96,9 +97,87 @@ describe("loadPublicSurveyPageUseCase", () => {
     vi.mocked(getPublicFormAccessUseCase).mockResolvedValue(
       Result.success(publicAccess),
     );
+    vi.mocked(getActiveDefinitionUseCase).mockResolvedValue(
+      Result.success(activeDefinition),
+    );
+    vi.mocked(getPartialSubmissionUseCase).mockResolvedValue(
+      ApiResult.notFoundError("No partial submission"),
+    );
   });
 
-  it("uses the access-token submission definition and skips active definition lookup", async () => {
+  it("loads access, submission, and definition in one async pass", async () => {
+    const definition = createDeferred<ResultType<ActiveDefinition>>();
+    vi.mocked(getActiveDefinitionUseCase).mockReturnValue(definition.promise);
+
+    const resultPromise = loadPublicSurveyPageUseCase({
+      formId,
+      tokenStore,
+    });
+
+    expect(getPublicFormAccessUseCase).toHaveBeenCalledWith({ formId });
+    expect(getPartialSubmissionUseCase).toHaveBeenCalledWith({
+      formId,
+      tokenStore,
+    });
+    expect(getActiveDefinitionUseCase).toHaveBeenCalledWith({ formId });
+
+    definition.resolve(Result.success(activeDefinition));
+    const result = await resultPromise;
+
+    expect(result.kind).toBe("success");
+  });
+
+  it("returns success with blocked phase when respondent already submitted", async () => {
+    vi.mocked(getPublicFormAccessUseCase).mockResolvedValue(
+      Result.success({
+        ...publicAccess,
+        limitOnePerUser: true,
+        hasUserSubmitted: true,
+        canStartNewSubmission: false,
+      }),
+    );
+
+    const result = await loadPublicSurveyPageUseCase({
+      formId,
+      tokenStore,
+    });
+
+    expect(result).toMatchObject({
+      kind: "success",
+      submissionPhase: "blocked",
+      activeDefinition,
+    });
+  });
+
+  it("keeps a cookie draft active when respondent already submitted", async () => {
+    vi.mocked(getPublicFormAccessUseCase).mockResolvedValue(
+      Result.success({
+        ...publicAccess,
+        limitOnePerUser: true,
+        hasUserSubmitted: true,
+        canStartNewSubmission: false,
+      }),
+    );
+    vi.mocked(getPartialSubmissionUseCase).mockResolvedValue(
+      ApiResult.success({
+        id: "submission-1",
+        isComplete: false,
+      } as Submission),
+    );
+
+    const result = await loadPublicSurveyPageUseCase({
+      formId,
+      tokenStore,
+    });
+
+    expect(result).toMatchObject({
+      kind: "success",
+      submissionPhase: "active",
+      submission: { id: "submission-1" },
+    });
+  });
+
+  it("loads token submissions through the access token use case", async () => {
     vi.mocked(getSubmissionByAccessTokenUseCase).mockResolvedValue(
       Result.success(accessTokenSubmission),
     );
@@ -127,6 +206,50 @@ describe("loadPublicSurveyPageUseCase", () => {
     expect(getPartialSubmissionUseCase).not.toHaveBeenCalled();
   });
 
+  it.each([
+    ["invalid token", ERROR_CODE.INVALID_TOKEN],
+    ["submission token invalid", ERROR_CODE.SUBMISSION_TOKEN_INVALID],
+    ["access forbidden", ERROR_CODE.ACCESS_FORBIDDEN],
+    ["authentication required", ERROR_CODE.AUTHENTICATION_REQUIRED],
+    ["resource not found", ERROR_CODE.RESOURCE_NOT_FOUND],
+    ["unknown error", ERROR_CODE.UNKNOWN_ERROR],
+  ] as const)(
+    "propagates token submission error code %s",
+    async (_name, errorCode) => {
+      vi.mocked(getSubmissionByAccessTokenUseCase).mockResolvedValue(
+        Result.error("Failed to load submission", undefined, errorCode),
+      );
+
+      const result = await loadPublicSurveyPageUseCase({
+        formId,
+        tokenStore,
+        urlToken: token,
+      });
+
+      expect(result).toEqual({
+        kind: "tokenSubmissionError",
+        errorCode,
+      });
+    },
+  );
+
+  it("falls back to unknown_error when access token result has no error code", async () => {
+    vi.mocked(getSubmissionByAccessTokenUseCase).mockResolvedValue(
+      Result.error("Failed to load submission"),
+    );
+
+    const result = await loadPublicSurveyPageUseCase({
+      formId,
+      tokenStore,
+      urlToken: token,
+    });
+
+    expect(result).toEqual({
+      kind: "tokenSubmissionError",
+      errorCode: ERROR_CODE.UNKNOWN_ERROR,
+    });
+  });
+
   it("keeps the active definition lookup for non-token public form loads", async () => {
     vi.mocked(getPartialSubmissionUseCase).mockResolvedValue(
       ApiResult.notFoundError("No partial submission"),
@@ -149,4 +272,58 @@ describe("loadPublicSurveyPageUseCase", () => {
     expect(getActiveDefinitionUseCase).toHaveBeenCalledWith({ formId });
     expect(getSubmissionByAccessTokenUseCase).not.toHaveBeenCalled();
   });
+
+  it("propagates partial submission lookup failures instead of treating them as no draft", async () => {
+    vi.mocked(getPartialSubmissionUseCase).mockResolvedValue(
+      ApiResult.networkError("Upstream unavailable"),
+    );
+
+    const result = await loadPublicSurveyPageUseCase({
+      formId,
+      tokenStore,
+    });
+
+    expect(result).toEqual({
+      kind: "submissionLoadError",
+      errorCode: ERROR_CODE.NETWORK_ERROR,
+    });
+  });
+
+  it("returns notFound when access or definition cannot be loaded", async () => {
+    vi.mocked(getPublicFormAccessUseCase).mockResolvedValue(
+      Result.error("Access denied"),
+    );
+
+    const accessResult = await loadPublicSurveyPageUseCase({
+      formId,
+      tokenStore,
+    });
+
+    vi.mocked(getPublicFormAccessUseCase).mockResolvedValue(
+      Result.success(publicAccess),
+    );
+    vi.mocked(getActiveDefinitionUseCase).mockResolvedValue(
+      Result.error("Definition missing"),
+    );
+
+    const definitionResult = await loadPublicSurveyPageUseCase({
+      formId,
+      tokenStore,
+    });
+
+    expect(accessResult).toEqual({ kind: "notFound" });
+    expect(definitionResult).toEqual({ kind: "notFound" });
+  });
 });
+
+function createDeferred<T>(): {
+  promise: Promise<T>;
+  resolve: (value: T) => void;
+} {
+  let resolve: (value: T) => void = () => {};
+  const promise = new Promise<T>((promiseResolve) => {
+    resolve = promiseResolve;
+  });
+
+  return { promise, resolve };
+}
